Replace "case" with "with" statement in Elixir day 21
This is to improve the efficiency slightly. The previous case statement
requires both the left and right monkey's to be evaluated, even in the
case where the first evaluated is :human. Using the "with" statement,
this is no longer the case - it can shortcut the evaluation if the first
evaluated is :human.